working NeRF am very proud, thanks mom
