+++ configs/auto/ct_data_13_256_64_486.txt
@@ -1,6 +1,6 @@
-expname = {}
+expname = ct_data_13_256_64_486
 basedir = ./logs/auto
-datadir = ./export/{}
+datadir = ./export/ct_data_13_256_64_486
 dataset_type = blender
 
 no_batching = True
