+++ configs/auto/ct_data_16_256_32_754.txt
@@ -1,6 +1,6 @@
-expname = {}
+expname = ct_data_16_256_32_754
 basedir = ./logs/auto
-datadir = ./export/{}
+datadir = ./export/ct_data_16_256_32_754
 dataset_type = blender
 
 no_batching = True
